requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
azure-core==1.35.0
azure-identity==1.23.1
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blockbuster==1.5.25
boto3==1.39.7
botocore==1.39.7
Bottleneck==1.5.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.2.1
cloudpickle==3.1.1
coloredlogs==15.0.1
comm==0.2.2
cryptography==44.0.3
dataclasses-json==0.6.7
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.16
dotenv==0.9.9
durationpy==0.10
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==25.2.10
forbiddenfruit==0.1.4
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.104.0
google-cloud-bigquery==3.35.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.26.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jsonschema_rs==0.29.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kubernetes==33.1.0
langchain==0.3.26
langchain-anthropic==0.3.17
langchain-aws==0.2.28
langchain-chroma==0.2.4
langchain-community==0.3.27
langchain-core==0.3.69
langchain-experimental==0.3.4
langchain-google-vertexai==2.0.27
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langgraph==0.5.3
langgraph-api==0.2.86
langgraph-checkpoint==2.1.0
langgraph-checkpoint-sqlite==2.0.10
langgraph-cli==0.3.4
langgraph-prebuilt==0.5.2
langgraph-runtime-inmem==0.3.4
langgraph-sdk==0.1.73
langgraph-supervisor==0.0.27
langgraph-swarm==0.0.13
langsmith==0.4.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mmh3==5.1.0
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
multidict==6.6.3
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.4
notebook_shim==0.2.4
numexpr==2.11.0
numpy==2.3.1
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.97.0
openevals==0.1.0
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.11.0
ormsgpack==1.10.0
overrides==7.7.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
posthog==5.4.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
PyPika==0.48.9
pyppeteer==0.0.25
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.0
scikit-learn==1.7.0
scipy==1.16.0
Send2Trash==1.8.3
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlite-vec==0.1.6
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.47.1
structlog==25.4.0
sympy==1.14.0
tavily-python==0.7.10
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
truststore==0.10.1
typer==0.16.0
types-python-dateutil==2.9.0.20250708
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
validators==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
